scrapers/requirements.txt
requests>=2.31,<3
feedparser>=6.0,<7
python-dateutil>=2.8,<3
pytest>=7.4,<9
responses>=0.24,<1
